fix hall bug
When session was lost and recovered, online_mode gets restarted. This shouldn't be a serious issue but in the case when the online player is P2, online agent's player will be  initialized as P1 before getting updated to P2. In that short period, board displays P1's info to P2.